% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genewalk.R
\name{check_degree_distribution}
\alias{check_degree_distribution}
\title{Check the degree across different node types}
\usage{
check_degree_distribution(object)
}
\arguments{
\item{object}{The \code{GeneWalk} class, please see \code{\link[=GeneWalk]{GeneWalk()}}.}
}
\description{
Shows the degrees of the different edge types in the GeneWalk network. Lack
of PPI interactions for example will make it difficult for the method to
identify patterns.
}
