% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_ext.R
\name{checkDiffusionProfilesParams}
\alias{checkDiffusionProfilesParams}
\title{Check diffusion profile parameters}
\usage{
checkDiffusionProfilesParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the diffusion profile
parameters.
}
\keyword{internal}
