% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_ext.R
\name{checkGeneWalkDataTable}
\alias{checkGeneWalkDataTable}
\title{Check GeneWalk input data.table}
\usage{
checkGeneWalkDataTable(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking GeneWalkDataTable parameters.
}
