% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_ext.R
\name{checkGeneWalkGraphDt}
\alias{checkGeneWalkGraphDt}
\title{Check GeneWalk graph data.table}
\usage{
checkGeneWalkGraphDt(x)
}
\arguments{
\item{x}{The data.table to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the graph_dt parameter of a
GeneWalk object.
}
\keyword{internal}
