% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkMetapath2VecParams}
\alias{checkMetapath2VecParams}
\title{Check metapath2vec parameters}
\usage{
checkMetapath2VecParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_metapath2vec]{params_metapath2vec()}}.
}
\keyword{internal}
