% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_ext.R
\name{checkNode2Vec}
\alias{checkNode2Vec}
\title{Check node2vec parameters}
\usage{
checkNode2Vec(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the node2vec parameters.
}
