% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{checkNode2VecParams}
\alias{checkNode2VecParams}
\title{Check node2vec parameters}
\usage{
checkNode2VecParams(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} if the check was successful, otherwise a
checkmate-style error string.
}
\description{
Checkmate extension for the output of \code{\link[=params_genewalk]{params_genewalk()}} and
\code{\link[=params_node2vec]{params_node2vec()}}.
}
\keyword{internal}
