% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_ext.R
\name{checkPPIParams}
\alias{checkPPIParams}
\title{Check PPI parameters}
\usage{
checkPPIParams(x)
}
\arguments{
\item{x}{The list to check/assert.}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking PPI parameters.
}
