% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{db_path}
\alias{db_path}
\title{Get the path to the cached database, downloading if necessary}
\usage{
db_path()
}
\description{
Get the path to the cached database, downloading if necessary
}
\keyword{internal}
