% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{dim.DataBuilder}
\alias{dim.DataBuilder}
\title{Dimensions of a DataBuilder object}
\usage{
\method{dim}{DataBuilder}(x)
}
\arguments{
\item{x}{A \code{DataBuilder} object.}
}
\value{
An integer vector of length 2, where the first element is the total
number of edges across all tables and the second element is \code{NA}.
}
\description{
Dimensions of a DataBuilder object
}
\keyword{internal}
