% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.get_node_degree}
\alias{.get_node_degree}
\title{Helper function to get the node degree from the PPI}
\usage{
.get_node_degree(graph_dt)
}
\arguments{
\item{graph_dt}{data.table. The graph info for which to get the node degree
of the genes.}
}
\value{
data.table with the gene and node degree in the network
}
\description{
Helper function to get the node degree from the PPI
}
\keyword{internal}
