% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{.string_cols_to_factors}
\alias{.string_cols_to_factors}
\title{Transform string to factor columns}
\usage{
.string_cols_to_factors(table)
}
\arguments{
\item{table}{data.table. The data.table for which to transform the factor
columns into strings.}
}
\value{
The data.table with the factor columns transformed to strings.
}
\description{
Transform string to factor columns
}
