% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{download_data}
\alias{download_data}
\title{Helper function to download necessary data}
\usage{
download_data(target_path)
}
\arguments{
\item{target_path}{String. Where do download the data}
}
\value{
Downloads the necessary files to target_path.
}
\description{
Helper function to download necessary data
}
\keyword{internal}
