% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_database}
\alias{download_database}
\title{Download actual database from Git LFS}
\usage{
download_database(repo = "GregorLueg/genewalkR", branch = "main")
}
\arguments{
\item{repo}{String. Repository in format "user/repo" (default:
\code{"GregorLueg/genewalkR"}).}

\item{branch}{String. Branch name (default: \code{"main"})}
}
\description{
Download actual database from Git LFS
}
