% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{download_pathway_commons}
\alias{download_pathway_commons}
\title{Helper function to download the pathway commons data}
\usage{
download_pathway_commons(
  target_path,
  url = "https://download.baderlab.org/PathwayCommons/PC2/v14/pc-hgnc.txt.gz"
)
}
\arguments{
\item{target_path}{String. Where do download the data}
}
\value{
Downloads the necessary files to target_path.
}
\description{
Helper function to download the pathway commons data
}
\keyword{internal}
