% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{evaluate_node2vec_test}
\alias{evaluate_node2vec_test}
\title{Evaluate node2vec test results}
\usage{
evaluate_node2vec_test(embeddings, node_labels)
}
\arguments{
\item{embeddings}{Matrix of node embeddings (nodes x dimensions).}

\item{node_labels}{data.table with columns: node, cluster.}
}
\value{
A list with:
\itemize{
\item{within_cluster_sim}{Mean similarity within clusters}
\item{between_cluster_sim}{Mean similarity between clusters}
\item{within_dist}{The within cluster similarities as vector}
\item{between_dist}{The between cluster similarities as vector}
}
}
\description{
Compares embedding similarities against ground truth clusters.
}
