% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g2g_interactions}
\alias{g2g_interactions}
\title{G2G Interactions Network}
\format{
A list with two elements:
\describe{
\item{graph}{A \code{data.table} with columns: \code{from}, \code{to}, \code{from_intact}, \code{from_string}, \code{from_signor}, \code{from_reactome}, \code{edge_type}, \code{source}}
\item{node_id}{A \code{data.table} with columns: \code{id} (ENSEMBL gene ID), \code{name} (gene name), \code{node_type} (character "gene")}
}
}
\source{
OpenTargets
}
\usage{
g2g_interactions
}
\description{
A combined dataset of gene-to-gene interactions from OpenTargets.
This dataset includes any interaction present in any of the source resources:
IntAct, STRING, Signor, or Reactome.
}
\keyword{datasets}
