% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{generate_combined_network}
\alias{generate_combined_network}
\title{Combine}
\usage{
generate_combined_network(
  interactions_string,
  interactions_signor,
  interactions_reactome,
  interactions_intact,
  interactions_pc
)
}
\arguments{
\item{interactions_string}{data.table. STRING interactions.}

\item{interactions_signor}{data.table. SIGNOR interactions.}

\item{interactions_reactome}{data.table. Reactome interactions.}

\item{interactions_intact}{data.table. Intact interactions.}

\item{interactions_pc}{data.table. Pathway commons interactions.}
}
\value{
The data.table with the Intact interaction
}
\description{
Combine
}
\keyword{internal}
