% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genewalk.R
\name{generate_permuted_emb}
\alias{generate_permuted_emb}
\title{Generate the random, permuted embedding}
\usage{
generate_permuted_emb(object, n_perm = 3L, seed = 42L, .verbose = TRUE)
}
\arguments{
\item{object}{The \code{genewalkR_class} class, please see
\code{\link[=genewalkR_class]{genewalkR_class()}}.}

\item{n_perm}{Inter. Number of permutations to generate. Defaults to \code{3L},
the recommendation of the authors.}

\item{seed}{Integer. Seed for reproducibility.}

\item{.verbose}{Boolean. Controls verbosity of the function}
}
\value{
The object with the added random embeddings for statistical testing.
}
\description{
This function will generate generate permuted subgraphs that show the same
node degree and run the node2vec algorithm on top of these subgraphs. The
cosine distance between the real edges will be calculated and stored in
the object. These will be used subsequently for statistical significance
testing. For more details, please refer to Ietswaart, et al. The function
will return a warning and the class as is if you have not run
\code{\link[=generate_initial_emb]{generate_initial_emb()}}.
}
\details{
The function will be using the same parameters for the node2vec training
as were used for \code{\link[=generate_initial_emb]{generate_initial_emb()}}.
}
\references{
Ietswaart, et al., Genome Biol, 2021
}
