% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GeneWalk}
\alias{GeneWalk}
\title{GeneWalk}
\usage{
GeneWalk(graph_dt, gene_to_pathway_dt, gene_ids, pathway_ids)
}
\arguments{
\item{graph_dt}{A data.table with the graph information. Needs to have
the columns \code{"from"}, \code{"to"} and \code{"type"}; type must include \code{"hierarchy"}
for the pathways, \code{"interaction"} for interactions and \code{"part_of"} for
gene/pathway interactions, and optionally a weight.}

\item{gene_to_pathway_dt}{A data.table with the gene to pathway information.
Needs to have \code{"from"} (gene) and \code{"to"} (pathway).}

\item{gene_ids}{String. The "bag of genes" you are testing.}

\item{pathway_ids}{String. The pathways represented in the network.}
}
\value{
Returns the initialised GeneWalk initialised object for subsequent
analysis.
}
\description{
Class that stores keeps all of the important data for the Gene Walk
algorithm, see Ietswaart, et al. The class can be subsequently used for to
apply the Gene Walk algorithm.
}
\section{Properties}{

\describe{
\item{graph_dt}{A data.table that contains the initial graph.}
\item{gene_to_pathway_dt}{A data.table that contains the gene to pathway
connections.}
\item{gene_ids}{The genes represented in your gene walk list.}
\item{pathway_ids}{The pathways represented in your gene walk network.}
\item{embds}{The initial embedding matrices generated by node2vec on top of
the graph. Depends on the n_graph parameter you choose.}
\item{permuted_embd}{A list containing cosine distances of permuted
graphs..}
\item{stats}{A data.table containing the information on the statistical
testing across the permutations.}
\item{params}{A list in which used parameters will be stored.}
}
}

\references{
Ietswaart, et al., Genome Biol, 2021
}
