% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_aligned_embd}
\alias{get_aligned_embd}
\title{Get the aligned embedding}
\usage{
get_aligned_embd(object)
}
\arguments{
\item{object}{An \code{EmbedDrift} object, see \code{\link[=EmbedDrift]{EmbedDrift()}}.}
}
\value{
A numeric matrix of the Procrustes-aligned embedding of graph 1
onto graph 2, restricted to shared nodes.
}
\description{
Get the aligned embedding
}
