% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_db_connection}
\alias{get_db_connection}
\title{Internal connector to the DB}
\usage{
get_db_connection()
}
\value{
The DuckDB connection to the internal data
}
\description{
Internal connector to the DB
}
