% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_graph_dt}
\alias{get_default_graph_dt}
\title{Generate default graph_dt from the network datasets provided by the package}
\usage{
get_default_graph_dt(
  resources = c("g2g_interactions", "reactome", "gene_ontology"),
  g2g_source = c("all", "intact", "string", "signor", "reactome"),
  go_type = c("all", "BP", "MF", "CC")
)
}
\arguments{
\item{resources}{Character vector of networks to include.
Choices: "g2g_interactions", "reactome", "gene_ontology".}

\item{g2g_source}{Which G2G interaction sources to include.}

\item{go_type}{Gene Ontology namespace to include.}
}
\value{
A list with:
\describe{
\item{graph_dt}{A \code{data.table} suitable for \code{genewalkR_class}}
\item{graph_gene_params}{Metadata describing how the graph was generated}
}
}
\description{
Generate default graph_dt from the network datasets provided by the package
}
