% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_embedding}
\alias{get_embedding}
\title{Get the list of embeddings}
\usage{
get_embedding(object)
}
\arguments{
\item{object}{The \code{GeneWalk} class, please see \code{\link[=GeneWalk]{GeneWalk()}}.}
}
\value{
If found, returns the embedding list.
}
\description{
Get the list of embeddings
}
