% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_embeddings}
\alias{get_embeddings}
\title{Get node embeddings}
\usage{
get_embeddings(object, which = NULL)
}
\arguments{
\item{object}{An \code{EmbedDrift} object, see \code{\link[=EmbedDrift]{EmbedDrift()}}.}

\item{which}{Integer. Which embedding to return: \code{1}, \code{2}, or \code{NULL} for
both as a named list. Defaults to \code{NULL}.}
}
\value{
A matrix, or a named list with elements \code{embd_1} and \code{embd_2}.
}
\description{
Get node embeddings
}
