% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_gene_ontology_info}
\alias{get_gene_ontology_info}
\title{Get the Gene Ontology main information}
\usage{
get_gene_ontology_info(head_only = FALSE)
}
\arguments{
\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the Gene Ontology information
}
\description{
Get the Gene Ontology main information
}
\details{
The data is extracted from the obo file downloaded from
\href{https://obofoundry.org/ontology/go.html}{GO OBO Foundry}. Last
download: \verb{March '26}.
}
