% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_gene_pathway_dt}
\alias{get_gene_pathway_dt}
\title{Get the gene to pathway data.table}
\usage{
get_gene_pathway_dt(object)
}
\arguments{
\item{object}{The \code{GeneWalk} class, please see \code{\link[=GeneWalk]{GeneWalk()}}.}
}
\value{
Returns the gene to pathway data.table stored in the object
}
\description{
Get the gene to pathway data.table
}
