% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_graph_dt}
\alias{get_graph_dt}
\title{Get the edge data.table}
\usage{
get_graph_dt(object)
}
\arguments{
\item{object}{The \code{GeneWalk} class, please see \code{\link[=GeneWalk]{GeneWalk()}}.}
}
\value{
Returns the edge data.table stored in the object
}
\description{
Get the edge data.table
}
