% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_graph_dts}
\alias{get_graph_dts}
\title{Get graph edge tables}
\usage{
get_graph_dts(object, which = NULL)
}
\arguments{
\item{object}{An \code{EmbedDrift} object, see \code{\link[=EmbedDrift]{EmbedDrift()}}.}

\item{which}{Integer. Which graph table to return: \code{1}, \code{2}, or \code{NULL} for
both as a named list. Defaults to \code{NULL}.}
}
\value{
A data.frame, or a named list with elements \code{graph_dt_1} and
\code{graph_dt_2}.
}
\description{
Get graph edge tables
}
