% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_gw_data}
\alias{get_gw_data}
\alias{get_gw_data.DataBuilder}
\title{Get the full GW data stored in the class}
\usage{
get_gw_data(x)

\method{get_gw_data}{DataBuilder}(x)
}
\arguments{
\item{x}{An object from which to extract the genewalk data.}
}
\value{
A data.table with the full data stored in the class.
}
\description{
Get the full GW data stored in the class
}
