% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_gw_data_filtered}
\alias{get_gw_data_filtered}
\alias{get_gw_data_filtered.DataBuilder}
\title{Get the filtered GW data for a list of genes}
\usage{
get_gw_data_filtered(x, gene_ids)

\method{get_gw_data_filtered}{DataBuilder}(x, gene_ids)
}
\arguments{
\item{x}{An object from which to extract the genewalk data}

\item{gene_ids}{String. The gene ids to which to filter the data.}
}
\value{
A list with the following elements
\itemize{
\item gwn - The final GeneWalk network.
\item genes_to_pathways - The genes to pathway information.
\item represented_genes - All included genes.
\item represented_pathways - All included pathways.
}
}
\description{
Get the filtered GW data for a list of genes
}
