% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_interactions_intact}
\alias{get_interactions_intact}
\title{Get the Intact network}
\usage{
get_interactions_intact(
  threshold = NULL,
  physical_interactions = FALSE,
  head_only = FALSE
)
}
\arguments{
\item{threshold}{Optional numeric. If you wish to filter on the STRING score.}

\item{physical_interactions}{Boolean. Shall only physical interactions be
returned.}

\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the Intact network data.
}
\description{
Get the Intact network
}
\details{
The data is extracted from the OpenTargets release \code{26.03}
}
