% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_interactions_pc}
\alias{get_interactions_pc}
\title{Get the pathway commons network}
\usage{
get_interactions_pc(head_only = FALSE)
}
\arguments{
\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the Pathway Commons network data.
}
\description{
Get the pathway commons network
}
\details{
This is version 14 of Pathway Commons. For details, check
Rodchenkov, et al., Nucleic Acids Res., 2019
}
