% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_interactions_string}
\alias{get_interactions_string}
\title{Get the STRING network}
\usage{
get_interactions_string(threshold = NULL, head_only = FALSE)
}
\arguments{
\item{threshold}{Optional numeric. If you wish to filter on the STRING score.}

\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the String network data.
}
\description{
Get the STRING network
}
\details{
The data is extracted from the OpenTargets release \code{26.03}
}
