% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_profiles}
\alias{get_profiles}
\title{Get diffusion profiles}
\usage{
get_profiles(object)
}
\arguments{
\item{object}{An object containing diffusion profiles.}
}
\value{
A numeric matrix of nodes x seeds.
}
\description{
Get diffusion profiles
}
