% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_reactome_hierarchy}
\alias{get_reactome_hierarchy}
\title{Get the Reactome pathway hierarchy}
\usage{
get_reactome_hierarchy(
  relationship = c("child_of", "descendant_of", "parent_of", "ancestor_of"),
  head_only = FALSE
)
}
\arguments{
\item{relationship}{String vector. Defines which relationships to return.
Multiple allowed and choices from
\code{c("child_of", "descendant_of", "parent_of", "ancestor_of")}}

\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the Reactome to Reactome edges with edge type info.
}
\description{
Get the Reactome pathway hierarchy
}
\details{
The data is extracted from the OpenTargets release \code{26.03}
}
