% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_reactome_info}
\alias{get_reactome_info}
\title{Get the Reactome main information}
\usage{
get_reactome_info(head_only = FALSE)
}
\arguments{
\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the Reactome information
}
\description{
Get the Reactome main information
}
\details{
The data is extracted from the OpenTargets release \code{26.03}
}
