% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_generics.R
\name{get_stats}
\alias{get_stats}
\title{Get the statistical results}
\usage{
get_stats(object)
}
\arguments{
\item{object}{The \code{GeneWalk} (see \code{\link[=GeneWalk]{GeneWalk()}}) or \code{EmbedDrift} class (see
\code{\link[=EmbedDrift]{EmbedDrift()}}).}
}
\value{
If found, the statistics results within the class
}
\description{
Get the statistical results
}
