% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_gene_data}
\alias{ingest_gene_data}
\title{Generates a genes main table}
\usage{
ingest_gene_data(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
A data.table with the genes to gene information (Ensembl
identifiers, symbols and biotype)
}
\description{
Generates a genes main table
}
\keyword{internal}
