% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_gene_ontology_genes}
\alias{ingest_gene_ontology_genes}
\title{Generates a Gene Ontology to gene table}
\usage{
ingest_gene_ontology_genes(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
A data.table with the Genes to Gene Ontology term associations
}
\description{
Generates a Gene Ontology to gene table
}
\keyword{internal}
