% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_gene_ontology_relations}
\alias{ingest_gene_ontology_relations}
\title{Generates the Gene Ontology table}
\usage{
ingest_gene_ontology_relations(obo_list)
}
\arguments{
\item{obo_list}{Named list. Output of \code{parse_gene_ontology()}}
}
\value{
A data.table with the Gene Ontology relationships
}
\description{
Generates the Gene Ontology table
}
\keyword{internal}
