% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_go_data}
\alias{ingest_go_data}
\title{Generates the Gene Ontology table}
\usage{
ingest_go_data(obo_list)
}
\arguments{
\item{obo_list}{Named list. Output of \code{parse_gene_ontology()}}
}
\value{
A data.table with the Gene ontology term ids, names and their
namespace
}
\description{
Generates the Gene Ontology table
}
\keyword{internal}
