% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_intact_interaction}
\alias{ingest_intact_interaction}
\title{Pull out the Intact interaction data from OpenTargets}
\usage{
ingest_intact_interaction(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
The data.table with the Intact interaction
}
\description{
Pull out the Intact interaction data from OpenTargets
}
\keyword{internal}
