% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_pc_interactions}
\alias{ingest_pc_interactions}
\title{Pull out the Pathway Commons interaction from the data}
\usage{
ingest_pc_interactions(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
The data.table with the PC interaction
}
\description{
Pull out the Pathway Commons interaction from the data
}
\keyword{internal}
