% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_reactome_data}
\alias{ingest_reactome_data}
\title{Generates a Reactome main table}
\usage{
ingest_reactome_data(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
A data.table with the Reactome main table
}
\description{
Generates a Reactome main table
}
\keyword{internal}
