% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_reactome_genes}
\alias{ingest_reactome_genes}
\title{Generates a Reactome to gene table}
\usage{
ingest_reactome_genes(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
A data.table with the genes to Reactome associations
}
\description{
Generates a Reactome to gene table
}
\keyword{internal}
