% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_reactome_interactions}
\alias{ingest_reactome_interactions}
\title{Pull out the Reactome interaction data from OpenTargets}
\usage{
ingest_reactome_interactions(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
The data.table with the Reactome interaction
}
\description{
Pull out the Reactome interaction data from OpenTargets
}
\keyword{internal}
