% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_reactome_ontology}
\alias{ingest_reactome_ontology}
\title{Generates a Reactome term ontology table}
\usage{
ingest_reactome_ontology(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
A data.table with the Reactome ontology (child of, etc.)
}
\description{
Generates a Reactome term ontology table
}
\keyword{internal}
