% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_signor_interactions}
\alias{ingest_signor_interactions}
\title{Pull out the SIGNOR interaction data from OpenTargets}
\usage{
ingest_signor_interactions(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
The data.table with the SIGNOR interaction
}
\description{
Pull out the SIGNOR interaction data from OpenTargets
}
\keyword{internal}
