% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{ingest_string_interactions}
\alias{ingest_string_interactions}
\title{Pull out the string data from OpenTargets}
\usage{
ingest_string_interactions(dir_data, threshold = 0.85)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}

\item{threshold}{Numeric. The string threshold from which to consider a
STRING connection high quality.}
}
\value{
The data.table with the STRING network data.
}
\description{
Pull out the string data from OpenTargets
}
\keyword{internal}
