% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{is_lfs_pointer}
\alias{is_lfs_pointer}
\title{Check if database file is a Git LFS pointer}
\usage{
is_lfs_pointer(db_path)
}
\arguments{
\item{db_path}{String. The path to the DB}
}
\description{
Check if database file is a Git LFS pointer
}
\keyword{internal}
