% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myc_genes}
\alias{myc_genes}
\title{MYC target genes}
\format{
\subsection{\code{myc_genes}}{

A data.table with 200 rows and 3 columns:
\describe{
\item{ensembl_gene}{Ensembl identifier}
\item{gene_symbol}{Gene Symbol}
\item{gene_set}{Name of the gene set}
}
}
}
\usage{
myc_genes
}
\description{
The MYC target genes from the MSigDB (Hallmarks V1).
}
\references{
Liberzon, et al., Cell Sys, 2014
}
\keyword{datasets}
