% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{n_nodes}
\alias{n_nodes}
\alias{n_nodes.DataBuilder}
\title{Number of nodes represented}
\usage{
n_nodes(x)

\method{n_nodes}{DataBuilder}(x)
}
\arguments{
\item{x}{An object from which to get the number of nodes.}
}
\value{
Number of unique nodes stored in the class.
}
\description{
Number of nodes represented
}
