% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_data_builder}
\alias{new_data_builder}
\title{Generate a new DataBuilder helper class}
\usage{
new_data_builder(ppis, gene_to_pathways, pathway_hierarchy)
}
\arguments{
\item{ppis}{data.table. The full interaction data to subset subsequently.}

\item{gene_to_pathways}{data.table. The gene to pathway associations.}

\item{pathway_hierarchy}{data.table. The pathway ontology}
}
\value{
Returns the initialised \code{DataBuilder} object.
}
\description{
Generate a new DataBuilder helper class
}
