% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node2vec.R
\name{node2vec}
\alias{node2vec}
\title{Node2Vec}
\usage{
node2vec(
  graph_dt,
  embd_dim = 8L,
  node2vec_params = params_node2vec(),
  backend = c("ndarray", "tch-cpu"),
  directed = FALSE,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{graph_dt}{data.table. The edge data.frame. Needs to have the columns
\code{"from"} and \code{"to"}. Can also optionally a \code{"weight"} column.}

\item{embd_dim}{Integer. Size of the embedding dimensions to create. Defaults
to \code{8L} in line with the authors recommendations.}

\item{node2vec_params}{Named list. Contains the node2vec parameters. The
list has the following elements:
\itemize{
\item p - Numeric. Return parameter for biased random walks. Defaults to
\code{1.0}.
\item q - Numeric. In-out parameter for biased random walks. Defaults to
\verb{1.0.}.
\item walks_per_node - Integer. Number of random walks per node. Defaults to
\verb{ 40L.}.
\item walk_length - Integer. Length of each random walk. Defaults to \code{40L}.
\item num_workers - Number of worker threads during batching. Defaults to
\code{4L}. To note: during graph generation, Rust will use via the Rayon backend
all available threads.
\item batch_size - Integer. Batch size for training. Defaults to \code{256L}.
\item n_epochs - Integer. Number of training epochs. Defaults to \code{15L}.
\item n_negatives - Integer. Number of negative samples. Defaults to \code{5L}.
\item window_size - Integer.  Context window size. Defaults to \code{2L}.
\item lr - Numeric. Learning rate. Defaults to \code{1e-2}.
}}

\item{backend}{String. One of \code{c("ndarray", "tch-cpu")}. \code{"tch-cpu"} is
usually much faster and supported only on Unix-based systems.}

\item{directed}{Boolean. Indicates if this is a directed or undirected
network. Defaults to \code{FALSE}.}

\item{seed}{Integer. Seed for reproducibility.}

\item{.verbose}{Boolean. Controls verbosity of the function}
}
\description{
Runs the Node2Vec algorithm
}
