% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{obo_parser}
\alias{obo_parser}
\title{Parse OBO file to list of data.tables}
\usage{
obo_parser(file_path, verbose = TRUE)
}
\arguments{
\item{file_path}{String. Path to the .obo file}

\item{verbose}{Boolean. Controls verbosity of the function.}
}
\value{
A list the following elements:
\itemize{
\item names - data.table with name and ID of the term.
\item alt_ids - data.table with alternative ids.
\item namespace - data.table with the found name-spaces.
\item term_subsets - data.table with the found subsets.
\item term_obsolete - character vector of terms that were marked obsolete.
\item term_relations - data.table with the ontological relationships found
in the file.
\item synonyms - data.table with synonyms provided in the .obo file.
\item externalRef - data.table with external references
}
}
\description{
This is a generalised helper function to read in .obo files (ontology files)
and generate some useful format out of it.
}
\keyword{internal}
