% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_docs.R
\name{opentargets_source}
\alias{opentargets_source}
\title{OpenTargets Data Source}
\description{
OpenTargets Data Source
}
\details{
The data is extracted from the OpenTargets release \code{26.03}
}
\keyword{internal}
