% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_genewalk}
\alias{params_genewalk}
\title{Wrapper function for the GeneWalk-specific node2vec parameters}
\usage{
params_genewalk(
  p = 1,
  q = 1,
  walks_per_node = 100L,
  walk_length = 10L,
  batch_size = 256L,
  n_epochs = 5L,
  n_negatives = 5L,
  window_size = 1L,
  lr = 0.01,
  num_workers = 1L
)
}
\arguments{
\item{p}{Numeric. Return parameter for biased random walks. Defaults to
\code{1.0}.}

\item{q}{Numeric. In-out parameter for biased random walks. Defaults to
\code{1.0}.}

\item{walks_per_node}{Integer. Number of random walks per node. Defaults to
\code{100L}.}

\item{walk_length}{Integer. Length of each random walk. Defaults to \code{10L}.}

\item{batch_size}{Integer. Batch size for training. Defaults to \code{256L}.}

\item{n_epochs}{Integer. Number of training epochs. Defaults to \code{5L}.}

\item{n_negatives}{Integer. Number of negative samples. Defaults to \code{5L}.}

\item{window_size}{Integer. Context window size. For GeneWalk you want
embeddings from direct neighbours. Defaults to \code{1L}.}

\item{lr}{Numeric. Learning rate. Defaults to \code{0.01}.}

\item{num_workers}{Integer. Number of worker threads. Anything above \code{1L}
makes the fitting much faster, at the cost of reproducibility. Defaults to
\code{1L}.}
}
\value{
A named list with the following elements:
\itemize{
\item p - Numeric. Return parameter for biased random walks. Defaults to
\code{1.0}.
\item q - Numeric. In-out parameter for biased random walks. Defaults to
\code{1.0}.
\item walks_per_node - Integer. Number of random walks per node. Defaults to
\code{100L}.
\item walk_length - Integer. Length of each random walk. Defaults to \code{10L}.
\item num_workers - Integer. Number of worker threads. Anything above \code{1L}
makes the fitting much faster, at the cost of reproducibility. Defaults to
\code{1L}.
\item batch_size - Integer. Batch size for training. Defaults to \code{256L}.
\item n_epochs - Integer. Number of training epochs. Defaults to \code{5L}.
\item n_negatives - Integer. Number of negative samples. Defaults to \code{5L}.
\item window_size - Integer. Context window size. For GeneWalk you want
embeddings from direct neighbours. Defaults to \code{1L}.
\item lr - Numeric. Learning rate. Defaults to \code{0.01}.
}
}
\description{
Wrapper function for the GeneWalk-specific node2vec parameters
}
