% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params-generated.R
\name{params_genewalk_data}
\alias{params_genewalk_data}
\title{GeneWalk synthetic data params}
\usage{
params_genewalk_data(
  n_signal_genes = 150L,
  n_noise_genes = 150L,
  n_roots = 10L,
  depth = 3L,
  branching = 3L,
  p_lateral = 0.1,
  p_ppi = 0.05,
  min_annotations = 10L,
  max_annotations = 25L,
  min_noise_subtrees = 15L
)
}
\arguments{
\item{n_signal_genes}{Integer. Number of signal genes. Defaults to \code{150L}.}

\item{n_noise_genes}{Integer. Number of noise genes. Defaults to \code{150L}.}

\item{n_roots}{Integer. Number of ontology root nodes. Defaults to \code{10L}.}

\item{depth}{Integer. Depth of each ontology subtree. Defaults to \code{3L}.}

\item{branching}{Integer. Base branching factor per ontology node. Defaults
to \code{3L}.}

\item{p_lateral}{Numeric. Probability of a lateral edge between sibling
nodes. Defaults to \code{0.1}.}

\item{p_ppi}{Numeric. Probability of a PPI edge between any two genes in the
same group. Defaults to \code{0.05}.}

\item{min_annotations}{Integer. Minimum ontology term annotations per gene.
Defaults to \code{10L}.}

\item{max_annotations}{Integer. Maximum ontology term annotations per gene.
Defaults to \code{25L}.}

\item{min_noise_subtrees}{Integer. Minimum distinct ontology subtrees a noise
gene must span. Defaults to \code{15L}.}
}
\value{
A named list with the following elements:
\itemize{
\item n_signal_genes - Integer. Number of signal genes. Defaults to \code{150L}.
\item n_noise_genes - Integer. Number of noise genes. Defaults to \code{150L}.
\item n_roots - Integer. Number of ontology root nodes. Defaults to \code{10L}.
\item depth - Integer. Depth of each ontology subtree. Defaults to \code{3L}.
\item branching - Integer. Base branching factor per ontology node. Defaults
to \code{3L}.
\item p_lateral - Numeric. Probability of a lateral edge between sibling
nodes. Defaults to \code{0.1}.
\item p_ppi - Numeric. Probability of a PPI edge between any two genes in
the same group. Defaults to \code{0.05}.
\item min_annotations - Integer. Minimum ontology term annotations per gene.
Defaults to \code{10L}.
\item max_annotations - Integer. Maximum ontology term annotations per gene.
Defaults to \code{25L}.
\item min_noise_subtrees - Integer. Minimum distinct ontology subtrees a
noise gene must span. Defaults to \code{15L}.
}
}
\description{
GeneWalk synthetic data params
}
