% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_pathway}
\alias{params_pathway}
\title{Pathway parameters}
\usage{
params_pathway(
  n_pathways = 250L,
  pathway_depth = 4L,
  pathway_branching = 3L,
  n_focal_pathways = 15L,
  connections_per_gene = 3L
)
}
\arguments{
\item{n_pathways}{Integer. Number of pathways. Defaults to \code{250L}.}

\item{pathway_depth}{Integer. Maximum depth of pathway hierarchy. Defaults
to \code{4L}.}

\item{pathway_branching}{Integer. Average branching factor. Defaults to \code{3L}.}

\item{n_focal_pathways}{Integer. Number of focal pathways per community.
Defaults to \code{15L}.}

\item{connections_per_gene}{Integer. Number of pathway connections per gene.
Defaults to \code{3L}.}
}
\value{
A list with pathway parameters.
}
\description{
Pathway parameters
}
