% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_ppi}
\alias{params_ppi}
\title{PPI parameters}
\usage{
params_ppi(n_genes = 500L, ppi_m = 3L, min_community_size = 20L)
}
\arguments{
\item{n_genes}{Integer. Number of genes. Defaults to \code{500L}.}

\item{ppi_m}{Integer. Number of edges to attach in preferential attachment.
Defaults to \code{3L}.}

\item{min_community_size}{Integer. Minimum size for gene communities.
Defaults to \code{20L}.}
}
\value{
A list with PPI parameters.
}
\description{
PPI parameters
}
