% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{parse_gene_ontology}
\alias{parse_gene_ontology}
\title{Parses the Gene Ontology}
\usage{
parse_gene_ontology(dir_data)
}
\arguments{
\item{dir_data}{String. The directory in which you downloaded the data}
}
\value{
A named list with the different sub tables and data extracted from
the .obo file
}
\description{
Parses the Gene Ontology
}
\keyword{internal}
