% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_docs.R
\name{pathway_commons_source}
\alias{pathway_commons_source}
\title{Pathway Commons Data Source}
\description{
Pathway Commons Data Source
}
\details{
This is version 14 of Pathway Commons. For details, check
Rodchenkov, et al., Nucleic Acids Res., 2019
}
\keyword{internal}
