% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_individual_gene}
\alias{plot_individual_gene}
\title{Generate a plot for an individual gene}
\usage{
plot_individual_gene(object, gene_of_interest, top_n = 10L)
}
\arguments{
\item{object}{The \code{GeneWalk} class. The stats need to be available in the
object.}

\item{gene_of_interest}{String. The gene of interest for which to plot
the enriched pathways.}

\item{top_n}{Integer. Number of genes to plot}
}
\value{
ggplot2 object with the plot for the gene
}
\description{
Plots the enrichment for individual genes within your GeneWalk results.
This will plot the top_n enriched pathways for the gene of interest.
}
