% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_similarities}
\alias{plot_similarities}
\title{Plot similarity distributions}
\usage{
plot_similarities(object)
}
\arguments{
\item{object}{The \code{GeneWalk} class. Permutations and stats need to be
available in the object.}
}
\value{
A ggplot2 object with the similarity distributions.
}
\description{
Plots the cosine similarity distributions of the permuted embeddings
alongside the actual values as density histograms.
}
