% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactome}
\alias{reactome}
\title{Reactome Pathway Network}
\format{
A list with two elements:
\describe{
\item{graph}{A \code{data.table} of pathway relationships: \code{from}, \code{to}, \code{edge_type}, \code{source}}
\item{node_id}{A \code{data.table} with columns: \code{id} (Reactome pathway ID), \code{name} (pathway name)}
}
}
\source{
OpenTargets
}
\usage{
reactome
}
\description{
A dataset containing Reactome pathways and their relationships to other pathways and genes.
}
\keyword{datasets}
