% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_cosine_sim}
\alias{rs_cosine_sim}
\title{Cosine similarity between two vectors}
\usage{
rs_cosine_sim(a, b)
}
\arguments{
\item{a}{Numeric vector. Vector a for which to calculate the similarity.}

\item{b}{Numeric vector. Vector b for which to calculate the similarity.}
}
\value{
Cosine similarity between the two vectors
}
\description{
Rust function to quickly calculate Cosine distances.
}
