% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gene_walk_perm}
\alias{rs_gene_walk_perm}
\title{Generate permuted embeddings for null distribution}
\usage{
rs_gene_walk_perm(
  from,
  to,
  weights,
  gene_walk_params,
  n_gene_pathway_edges,
  n_perm,
  embd_dim,
  directed,
  seed,
  verbose
)
}
\arguments{
\item{from}{Integer vector. Node indices for edge origins.}

\item{to}{Integer vector. Node indices for edge destinations.}

\item{weights}{Optional numeric vector. Edge weights, defaults to 1.0.}

\item{gene_walk_params}{Named list. Training parameters.}

\item{n_perm}{Integer. Number of permutations.}

\item{embd_dim}{Integer. Embedding dimension.}

\item{directed}{Boolean. Treat graph as directed.}

\item{seed}{Integer. Random seed.}

\item{verbose}{Boolean. Controls verbosity.}
}
\value{
A list of n_perm cosine similarities based on the connected nodes
of node-degree matched random graphs.
}
\description{
Generates degree-preserving random networks and trains node2vec
on each, returning the raw embedding matrices for downstream statistical
testing.
}
