% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_kernel_node_names}
\alias{rs_kernel_node_names}
\title{Pull out the node names from the kernel}
\usage{
rs_kernel_node_names(kernel)
}
\arguments{
\item{kernel}{External pointer to the \code{DiffusionKernel}.}
}
\value{
A string vector from R.
}
\description{
Pull out the node names from the kernel
}
