% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_node2vec}
\alias{rs_node2vec}
\title{Generate a node2vec embeddings}
\usage{
rs_node2vec(
  from,
  to,
  weights,
  node2vec_params,
  embd_dim,
  directed,
  seed,
  verbose
)
}
\arguments{
\item{from}{Integer vector. Node indices for edge origins.}

\item{to}{Integer vector. Node indices for edge destinations.}

\item{weights}{Optional numeric vector. Edge weights, defaults to 1.0.}

\item{node2vec_params}{Named list. Training parameters (p, q,
walks_per_node, walk_length, num_workers, n_epochs, num_negatives,
window_size, lr, dim).}

\item{embd_dim}{Integer. Embedding dimension.}

\item{directed}{Boolean. Treat graph as directed.}

\item{seed}{Integer. Random seed (incremented per rep).}

\item{verbose}{Boolean. Controls verbosity.}
}
\value{
An embedding of dimension n_nodes x embedding dim.
}
\description{
Trains node2vec on the original provided graph.
}
