% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_node2vec_synthetic_data}
\alias{rs_node2vec_synthetic_data}
\title{Generate synthetic data for node2vec}
\usage{
rs_node2vec_synthetic_data(
  test_data,
  n_nodes_per_cluster,
  n_clusters,
  p_within,
  p_between,
  seed
)
}
\arguments{
\item{n_nodes_per_cluster}{Integer. Number of nodes in the test graph.}

\item{n_clusters}{Integer. Number of nodes per cluster.}

\item{p_within}{Numeric. Probability of edges within cluster (0-1).}

\item{p_between}{Numeric. Probability of edges between clusters (0-1).}

\item{seed}{Integer. Random seed.}

\item{String.}{One of \code{c("barbell", "caveman", "stochastic_block")}. Weird
strings will default to "barbell" data.}
}
\value{
A list with the following elements:
\itemize{
\item edges - List with the edge data.
\item nodes - List with the node information.
}
}
\description{
Generate synthetic data for node2vec
}
