% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_graph}
\alias{trim_graph}
\title{Trim a heterogeneous graph by gene set}
\usage{
trim_graph(graph_dt, node_id, genes_to_keep, gene_type = "gene")
}
\arguments{
\item{graph_dt}{data.table with at least \code{from} and \code{to}}

\item{node_id}{data.table with columns \code{id} and \code{type}}

\item{genes_to_keep}{character vector of gene IDs}

\item{gene_type}{value in \code{node_id$node_type} identifying genes}
}
\value{
list with trimmed \code{graph_dt} and \code{node_id}
}
\description{
Trim a heterogeneous graph by gene set
}
