% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_ext.R
\name{assertGeneWalkGraphDt}
\alias{assertGeneWalkGraphDt}
\title{Assert GeneWalk graph data.table}
\usage{
assertGeneWalkGraphDt(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The data.table to check.}

\item{.var.name}{Name of the checked object to print in assertions. Defaults
to the heuristic implemented in checkmate.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:AssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns the checked object if the assertion is successful.
}
\description{
Checkmate extension for asserting the graph_dt parameter of a
GeneWalk object.
}
\keyword{internal}
