% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-params-generated.R
\name{assertMetapath2VecParams}
\alias{assertMetapath2VecParams}
\title{Assert metapath2vec parameters}
\usage{
assertMetapath2VecParams(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{The object to check.}

\item{.var.name}{Name of the checked object to print in assertions.}

\item{add}{Collection to store assertion messages. See
\code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}.}
}
\value{
Invisibly returns the checked object if the assertion is
successful.
}
\description{
Assert metapath2vec parameters
}
\keyword{internal}
