% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_ext.R
\name{checkMetapath2VecParams}
\alias{checkMetapath2VecParams}
\title{Check metapath2vec parameters}
\usage{
checkMetapath2VecParams(x)
}
\arguments{
\item{x}{The list to check/assert}
}
\value{
\code{TRUE} if the check was successful, otherwise an error message.
}
\description{
Checkmate extension for checking the metapath2vec parameters.
}
\keyword{internal}
