% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_docs.R
\name{gene_ontology_source}
\alias{gene_ontology_source}
\title{Gene Ontology Data Source}
\description{
Gene Ontology Data Source
}
\details{
The data is extracted from the obo file downloaded from
\href{https://obofoundry.org/ontology/go.html}{GO OBO Foundry}. Last
download: \verb{January '26}.
}
\keyword{internal}
