% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genewalk.R
\name{generate_initial_emb}
\alias{generate_initial_emb}
\title{Generate the initial graph embedding}
\usage{
generate_initial_emb(
  object,
  embd_dim = 8L,
  node2vec_params = params_node2vec(),
  directed = FALSE,
  seed = 42L,
  .verbose = TRUE
)
}
\arguments{
\item{object}{The \code{genewalkR_class} class, please see
\code{\link[=genewalkR_class]{genewalkR_class()}}.}

\item{embd_dim}{Integer. Size of the embedding dimensions to create. Defaults
to \code{8L} in line with the authors recommendations.}

\item{node2vec_params}{Named list. Contains the node2vec parameters. The
list has the following elements:
\itemize{
\item p - Numeric. Return parameter for biased random walks. Defaults to
\code{1.0}.
\item q - Numeric. In-out parameter for biased random walks. Defaults to
\verb{1.0.}.
\item walks_per_node - Integer. Number of random walks per node. Defaults to
\verb{ 20L.}.
\item walk_length - Integer. Length of each random walk. Defaults to \code{20L}.
\item num_workers - Number of worker threads during batching. Defaults to
\code{4L}. To note: during graph generation, Rust will use via the Rayon backend
all available threads.
\item batch_size - Integer. Batch size for training. Defaults to \code{256L}.
\item num_epochs - Integer. Number of training epochs. Defaults to \code{5L}.
\item num_negatives - Integer. Number of negative samples. Defaults to \code{5L}.
\item window_size - Integer.  Context window size. Defaults to \code{2L}.
\item lr - Numeric. Learning rate. Defaults to \code{1e-3}.
}}

\item{directed}{Boolean. Indicates if this is a directed or undirected
network. Defaults to \code{FALSE}.}

\item{seed}{Integer. Seed for reproducibility.}

\item{.verbose}{Boolean. Controls verbosity of the function}
}
\value{
The object with the added initial embeddings.
}
\description{
This function will generate the initial node2vec embedding with the provided
parameters. In brief, it will generate random walks over the graph and fit
a Skip Gram model for the number of defined epochs to generate on a per node
basis an embedding vector. These will be subsequently used for the
calculation of test statistics. For more details, please refer to Ietswaart,
et al.
}
\references{
Ietswaart, et al., Genome Biol, 2021
}
