% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GeneWalkGenerator}
\alias{GeneWalkGenerator}
\title{Gene Walk Network Generator}
\description{
R6 class to build Gene Walk networks from internal database sources.
Supports building a master network once and subsetting to multiple gene sets.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pathway_sources}}{Character vector of pathway sources}

\item{\code{pathway_namespaces}}{List of namespace filters per pathway source}

\item{\code{ppi_sources}}{Character vector of PPI sources}

\item{\code{ppi_params}}{List of PPI filtering parameters}

\item{\code{network_dt}}{data.table containing the full network}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GeneWalkGenerator-new}{\code{GeneWalkGenerator$new()}}
\item \href{#method-GeneWalkGenerator-add_pathways}{\code{GeneWalkGenerator$add_pathways()}}
\item \href{#method-GeneWalkGenerator-add_ppi}{\code{GeneWalkGenerator$add_ppi()}}
\item \href{#method-GeneWalkGenerator-build}{\code{GeneWalkGenerator$build()}}
\item \href{#method-GeneWalkGenerator-create_for_genes}{\code{GeneWalkGenerator$create_for_genes()}}
\item \href{#method-GeneWalkGenerator-reset_choices}{\code{GeneWalkGenerator$reset_choices()}}
\item \href{#method-GeneWalkGenerator-return_full_network_dt}{\code{GeneWalkGenerator$return_full_network_dt()}}
\item \href{#method-GeneWalkGenerator-clone}{\code{GeneWalkGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneWalkGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-GeneWalkGenerator-new}{}}}
\subsection{Method \code{new()}}{
Initialise the generator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneWalkGenerator$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneWalkGenerator-add_pathways"></a>}}
\if{latex}{\out{\hypertarget{method-GeneWalkGenerator-add_pathways}{}}}
\subsection{Method \code{add_pathways()}}{
Add pathway sources to the network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneWalkGenerator$add_pathways(
  source = c("go", "reactome"),
  go_namespace = c("biological_process", "molecular_function", "cellular_component")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Character vector of sources: "go", "reactome"}

\item{\code{go_namespace}}{Character vector. Specifically for GO. Choices
are \code{c("biological_process", "molecular_function", "cellular_component")}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneWalkGenerator-add_ppi"></a>}}
\if{latex}{\out{\hypertarget{method-GeneWalkGenerator-add_ppi}{}}}
\subsection{Method \code{add_ppi()}}{
Add PPI sources to the network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneWalkGenerator$add_ppi(
  source = c("combined", "string", "signor", "reactome", "intact"),
  string_threshold = NULL,
  intact_threshold = NULL,
  intact_physical_only = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Character vector of sources}

\item{\code{string_threshold}}{Numeric threshold for STRING scores}

\item{\code{intact_threshold}}{Numeric threshold for Intact scores}

\item{\code{intact_physical_only}}{Logical, only physical interactions from Intact}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneWalkGenerator-build"></a>}}
\if{latex}{\out{\hypertarget{method-GeneWalkGenerator-build}{}}}
\subsection{Method \code{build()}}{
Build the full network from selected sources
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneWalkGenerator$build()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneWalkGenerator-create_for_genes"></a>}}
\if{latex}{\out{\hypertarget{method-GeneWalkGenerator-create_for_genes}{}}}
\subsection{Method \code{create_for_genes()}}{
Create a gene-specific genewalkR_class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneWalkGenerator$create_for_genes(genes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Character vector of gene symbols}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneWalkGenerator-reset_choices"></a>}}
\if{latex}{\out{\hypertarget{method-GeneWalkGenerator-reset_choices}{}}}
\subsection{Method \code{reset_choices()}}{
Resets the internal choices and erases any stored network.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneWalkGenerator$reset_choices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneWalkGenerator-return_full_network_dt"></a>}}
\if{latex}{\out{\hypertarget{method-GeneWalkGenerator-return_full_network_dt}{}}}
\subsection{Method \code{return_full_network_dt()}}{
Resets the internal choices and erases any stored network.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneWalkGenerator$return_full_network_dt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GeneWalkGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GeneWalkGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GeneWalkGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
