% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{genewalkR_class}
\alias{genewalkR_class}
\title{genewalkR}
\usage{
genewalkR_class(graph_dt, graph_gene_params)
}
\value{
Returns the initialised genewalkR_class object for subsequent
analysis.
}
\description{
Class that stores keeps all of the important data for the Gene Walk
algorithm, see Ietswaart, et al. The class can be subsequently used for to
apply the Gene Walk algorithm.
}
\section{Properties}{

\describe{
\item{graph_dt}{A data.table that contains the initial graph.}
\item{embd}{The initial embedding matrix generated by node2vec on top of
the graph.}
\item{permuted_embd}{A list containing the permuted versions of the graph
data for subsequent statistical testing.}
\item{global_stats}{A data.table containing the global statistics with the
FDR correction applied across all gene-pathway pairs.}
\item{gene_stats}{A data.table containing the gene-based statistics with
the FDR correction applied within a given gene.}
\item{params}{A (nested) list that will store all the parameters of the
applied function.}
}
}

\references{
Ietswaart, et al., Genome Biol, 2021
}
