% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_gene_info}
\alias{get_gene_info}
\title{Get the gene main information}
\usage{
get_gene_info(head_only = FALSE)
}
\arguments{
\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the gene information
}
\description{
Get the gene main information
}
\details{
The data is extracted from the OpenTargets release \code{25.12}
}
