% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_gene_ontology_hierarchy}
\alias{get_gene_ontology_hierarchy}
\title{Get the Gene Ontology}
\usage{
get_gene_ontology_hierarchy(head_only = FALSE)
}
\arguments{
\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the GO to GO edges with edge type info.
}
\description{
Get the Gene Ontology
}
\details{
The data is extracted from the obo file downloaded from
\href{https://obofoundry.org/ontology/go.html}{GO OBO Foundry}. Last
download: \verb{January '26}.
}
