% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_gene_to_reactome}
\alias{get_gene_to_reactome}
\title{Get the Reactome to gene connections}
\usage{
get_gene_to_reactome(head_only = FALSE)
}
\arguments{
\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the Reactome to gene edges
}
\description{
Get the Reactome to gene connections
}
\details{
The data is extracted from the OpenTargets release \code{25.12}
}
