% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_interactions_combined}
\alias{get_interactions_combined}
\title{Get the Combined network}
\usage{
get_interactions_combined(head_only = FALSE)
}
\arguments{
\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the Reactome network data.
}
\description{
This returns a combined network akin to the workflow from
Barrio-Hernandez et al. based on high quality edges from STRING, physical
interactions from Intact and the full sets from SIGNOR and Reactome
}
\details{
The data is extracted from the OpenTargets release \code{25.12}
}
\references{
Barrio-Hernandez et al., Nat Gen, 2023
}
