% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_interactions_reactome}
\alias{get_interactions_reactome}
\title{Get the Reactome network}
\usage{
get_interactions_reactome(head_only = FALSE)
}
\arguments{
\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the Reactome network data.
}
\description{
Get the Reactome network
}
\details{
The data is extracted from the OpenTargets release \code{25.12}
}
