% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_interactions_signor}
\alias{get_interactions_signor}
\title{Get the SIGNOR network}
\usage{
get_interactions_signor(head_only = FALSE)
}
\arguments{
\item{head_only}{Logical. If TRUE, returns only the first 5 rows.}
}
\value{
data.table with the SIGNOR network data.
}
\description{
Get the SIGNOR network
}
\details{
The data is extracted from the OpenTargets release \code{25.12}
}
