% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{get_stats}
\alias{get_stats}
\title{Get the statistical results}
\usage{
get_stats(object)
}
\arguments{
\item{object}{The \code{genewalkR_class} class, please see
\code{\link[=genewalkR_class]{genewalkR_class()}}.}
}
\value{
If found, the GeneWalk statistics results
}
\description{
Get the statistical results
}
