% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_node2vec}
\alias{params_node2vec}
\title{Wrapper function for the node2vec parameters}
\usage{
params_node2vec(
  p = 1,
  q = 1,
  walks_per_node = 40L,
  walk_length = 40L,
  batch_size = 256L,
  n_epochs = 20L,
  n_negatives = 5L,
  window_size = 2L,
  lr = 0.01,
  num_workers = NULL
)
}
\arguments{
\item{p}{Numeric. Return parameter for biased random walks. Defaults to
\code{1.0}.}

\item{q}{Numeric. In-out parameter for biased random walks. Defaults to
\code{1.0}.}

\item{walks_per_node}{Integer. Number of random walks per node. Defaults to
\code{40L}.}

\item{walk_length}{Integer. Length of each random walk. Defaults to \code{40L}.}

\item{batch_size}{Integer. Batch size for training. Defaults to \code{256L}.}

\item{n_epochs}{Integer. Number of training epochs. Defaults to \code{20L}.}

\item{n_negatives}{Integer. Number of negative samples. Defaults to \code{5L}.}

\item{window_size}{Integer. Context window size. Defaults to \code{2L}.}

\item{lr}{Numeric. Learning rate. Defaults to \code{1e-2}.}

\item{num_workers}{Optional integer. If kept to \code{NULL}, it will default
to \verb{available cores - 2 (min 1)}.}
}
\value{
A list with the node2vec parameters.
}
\description{
These are general node2vec parameters, designed for speed. They will default
to \verb{available cores - 2 (min 1)}.
}
