% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gw_results}
\alias{plot_gw_results}
\title{Plot the gene walk results}
\usage{
plot_gw_results(object)
}
\arguments{
\item{object}{The \code{GeneWalk} class. The stats need to be available in the
object.}
}
\value{
ggplot2 object with the plot for the gene
}
\description{
Plots the GeneWalk results. Creates an interactive plotly plot to explore
the results interactively.
}
