% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.EmbedDrift}
\alias{print.EmbedDrift}
\title{print Method for EmbedDrift object}
\arguments{
\item{x}{An object of class \code{EmbedDrift}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an EmbedDrift object.
}
\keyword{internal}
