% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.GeneWalk}
\alias{print.GeneWalk}
\title{print Method for GeneWalk object}
\arguments{
\item{x}{An object of class GeneWalk}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns x.
}
\description{
Print a GeneWalk object.
}
\keyword{internal}
