% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_diffuse}
\alias{rs_diffuse}
\title{Run diffusion scoring on a precomputed kernel}
\usage{
rs_diffuse(
  kernel,
  scores,
  n_bkgd,
  n_inputs,
  bkgd_indices,
  method,
  n_perm,
  seed
)
}
\arguments{
\item{kernel}{External pointer to \code{DiffusionKernel}}

\item{scores}{Numeric. The scores.}

\item{n_bkgd}{Integer. Number of background nodes.}

\item{n_inputs}{Integer. Number of input columns}

\item{bkgd_indices}{Integer. 0-based node indices for the background}

\item{method}{String. One of \code{c("raw", "z", "mc")}.}

\item{n_perm}{Integer. Number of permutations (MC only)}

\item{seed}{Integer RNG seed (MC only)}
}
\value{
Scores per given node
}
\description{
Run diffusion scoring on a precomputed kernel
}
