% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gene_walk}
\alias{rs_gene_walk}
\title{Function that generates the GeneWalk embedding}
\usage{
rs_gene_walk(
  from,
  to,
  weights,
  gene_walk_params,
  backend,
  embd_dim,
  directed,
  seed,
  verbose
)
}
\arguments{
\item{from}{Integer vector. The node indices indicating where the edge
originates from.}

\item{to}{Integer vector. The node indices indicating where the edge
goes to.}

\item{weights}{Optional numeric vector. If not supplied, defaults to \code{1.0}
as edge weight.}

\item{gene_walk_params}{Named list. Contains the parameters for running
gene walk.}

\item{backend}{String. One of \code{c("tch-cpu", "ndarray")}. Torch is not
supported on Windows and will panic.}

\item{embd_dim}{Integer. Embedding dimension.}

\item{directed}{Boolean. Is the graph directed. If set to \code{FALSE} reverse
edges will be added.}

\item{seed}{Integer. For reproducibility of the algorithm.}

\item{verbose}{Boolean. Controls verbosity of the algorithm.}
}
\value{
A numerical matrix of number nodes x desired embedding dimensios.
}
\description{
Wrapper function that leverages Rust and the Burn Tensor
framework to generate node representations for subsequent usage in the
GeneWalk approach. The default back end is the torch CPU backend. In the
future there might be other backends added, but testing revealed this one
to be very performant.
}
