% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gene_walk_perm}
\alias{rs_gene_walk_perm}
\title{Function to generate permuted embeddings}
\usage{
rs_gene_walk_perm(
  from,
  to,
  weights,
  gene_walk_params,
  n_perm,
  embd_dim,
  directed,
  seed,
  verbose
)
}
\arguments{
\item{from}{Integer vector. The node indices indicating where the edge
originates from.}

\item{to}{Integer vector. The node indices indicating where the edge
goes to.}

\item{weights}{Optional numeric vector. If not supplied, defaults to \code{1.0}
as edge weight.}

\item{gene_walk_params}{Named list. Contains the parameters for running
gene walk.}

\item{n_perm}{Integer.}

\item{embd_dim}{Integer. Embedding dimension.}

\item{directed}{Boolean. Is the graph directed. If set to \code{FALSE} reverse
edges will be added.}

\item{seed}{Integer. For reproducibility of the algorithm.}

\item{verbose}{Boolean. Controls verbosity of the algorithm.}
}
\value{
A list with permuted Cosine similarities between the edges.
}
\description{
Wrapper function to generate permuted embeddings and subsequently the
null distribution of cosine similarities for statistical testing.
}
