% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gene_walk_test}
\alias{rs_gene_walk_test}
\title{Calculate the test statistics}
\usage{
rs_gene_walk_test(
  gene_embds,
  pathway_embds,
  null_distribution,
  return_similarities,
  verbose
)
}
\arguments{
\item{gene_embds}{Matrix of n_genes x their graph embeddings}

\item{pathway_embds}{Matrix of n_pathways x their graph embeddings}

\item{null_distribution}{Numeric vector representing the null distributions
in terms of Cosine distances.}

\item{return_similarities}{Boolean. Shall the Cosine similarity matrix
between genes x pathways be returned}

\item{verbose}{Controls verbosity of the function.}
}
\value{
A list with
\itemize{
\item pvals - The calculated p-values for the genes x pathways based
on the permuted data.
\item sim - NULL if return_similarities=FALSE or the cosine similarities
in the embedding space.
}
}
\description{
Calculates the test statistic for the gene / pathway pairs.
}
