% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_gene_walk_test}
\alias{rs_gene_walk_test}
\title{Calculate the test statistics}
\usage{
rs_gene_walk_test(gene_embds, pathway_embds, null_distributions, verbose)
}
\arguments{
\item{gene_embds}{Matrix of n_genes x their graph embeddings}

\item{pathway_embds}{Matrix of n_pathways x their graph embeddings}

\item{null_distributions}{List of null distribution vectors (one per
permutation).}

\item{verbose}{Controls verbosity of the function.}
}
\value{
A list with vectors:
\itemize{
\item gene - Gene indices (1-based for R)
\item pathway - Pathway indices (1-based for R)
\item pval - P-values for each gene-pathway pair
\item global_fdr - FDR corrected across all pairs
\item gene_fdr - FDR corrected within each gene
}
}
\description{
Calculates the test statistic for the gene / pathway pairs.
}
