% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_node2vec_synthetic_data}
\alias{rs_node2vec_synthetic_data}
\title{Generate synthetic data for node2vec}
\usage{
rs_node2vec_synthetic_data(
  test_data,
  n_nodes_per_cluster,
  n_clusters,
  p_within,
  p_between,
  seed
)
}
\arguments{
\item{test_data}{String. One of
c("barbell", "caveman", "stochastic_block").}

\item{n_nodes_per_cluster}{Integer. Nodes per cluster.}

\item{n_clusters}{Integer. Number of clusters.}

\item{p_within}{Numeric. Within-cluster edge probability (0-1).}

\item{p_between}{Numeric. Between-cluster edge probability (0-1).}

\item{seed}{Integer. Random seed.}
}
\value{
A list with edges and nodes.
}
\description{
Generate synthetic data for node2vec
}
